requirements.txt
joblib==1.3.2
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.12.2
tqdm==4.66.4
wfdb==4.1.2
torch==2.1.1
torchvision==0.16.1
opencv-contrib-python==4.10.0.84